+++ HISTORY.rst
@@ -16,6 +16,7 @@ latest
 * Add a `diagnostics_utils` poetry package in `./external` and a `training_data_diagnostics` workflow step for processing
 training data from multiple sources into a set of diagnostic variables
 * Add a `diagnostic_sequence_from_mapper` to the `loaders` module function for loading data for diagnostics
+* Add nudging data loaders for use in diagnostics and ML training
 
 
 0.2.1 (2020-05-15)